refactor: refactoring utils; feat: adding document management class
